+++ src/app/experience/page.tsx
@@ -5,8 +5,8 @@ import Period from '@/components/ui/projects/Period/Period';
 
 import { EXPERIENCE_INFO } from '@/constants/experience';
 
-import type { Project as ProjectType } from '@/common/types/experience';
-import { Experience } from '@/common/types/experience';
+import type { Project as ProjectType } from '@/types/experience';
+import { Experience } from '@/types/experience';
 
 import type { NextPage } from 'next';
