+++ src/components/common/menus/MenuList/MenuList.tsx
@@ -2,14 +2,13 @@ import type { FC } from 'react';
 
 import { motion } from 'framer-motion';
 
-
 import { MENU_ITEMS } from '@/constants/menuItems';
 import { SOCIAL_MEDIA_LIST } from '@/constants/socialMedia';
 
 import getIconComponent from '@/utils/getIconComponent';
 
-import type { MenuItem } from '@/common/types/menuItems';
-import type { SocialMedia } from '@/common/types/socialMedia';
+import type { MenuItem } from '@/types/menuItems';
+import type { SocialMedia } from '@/types/socialMedia';
 
 const activeMenuVariants = {
   initial: {
