+++ src/constants/menuItems.ts
@@ -1,4 +1,4 @@
-import type { MenuItem } from '../common/types/menuItems';
+import type { MenuItem } from '../types/menuItems';
 
 export const MENU_ITEMS: MenuItem[] = [
   {
